[CMake] Fix usage of binary dir. Clean up and improve documentation.
Using CMAKE_CURRENT_BINARY_DIR now instead of MeddySDK_Meddyproject_BINARY_DIR. Because it's the same variable used in CMake's functions so we just want to ensure we're consitent.

Separate target_sources() into two so we have a separate one for headers.

Get rid of the "/headers" in the generated header include directory.