[CMake] [Tests] Fix configuration error when building project as top-level (e.g., outside of the MeddySDK superproject).
The error was because we were using `find_package` on MeddySDK_Meddyproject, but there was no FetchContent declaration of it overriding `find_package`, so it was failing to find it.

An alternative (and, arguably better) way to fix this is to structure both the test and non-test versions of the project into two different subdirectories, tied together by `FetchContent` from a parent project. This would allow one to find the other in a decouple manner. For now, we are just using `add_subdirectory`, though.

Document why we are okay with using `add_subdirectory` instead of `FetchContent` for our test subproject.